ngatives feedback

+++ ProsDataBase/database/views/api.py
@@ -32,26 +32,36 @@ def user(request, name):
 
 
 def userRights(request):
-    if request.method == 'GET':
-        return showUserRights(request)
-    if request.method == 'POST':
-        return modifyUserRights(request)
+    if request.user.userManager or request.user.admin :
+        if request.method == 'GET':
+            return showUserRights(request)
+        if request.method == 'POST':
+            return modifyUserRights(request)
+    else:
+        return HttpResponse('{"errors":[{"message":"'+(_("You have not the rights to do this opperation").__unicode__())+'"}]}',content_type="application/json")
 
 
 def groups(request):
     if request.method == 'GET':
         return showAllGroups()
     elif request.method == 'POST':
-        return addGroup(request)
+        if request.user.userManager or request.user.admin :
+            return addGroup(request)
+        else:
+            return HttpResponse('{"errors":[{"message":"'+(_("You have not the rights to do this opperation").__unicode__())+'"}]}',content_type="application/json")
+
 
 
 def group(request, name):
     if request.method == 'GET':
         return showOneGroup(name)
-    if request.method == 'PUT':
-        return modifyGroup(request, name)
-    if request.method == 'DELETE':
-        return deleteGroup(request, name)
+    if request.user.userManager or request.user.admin :
+        if request.method == 'PUT':
+            return modifyGroup(request, name)
+        if request.method == 'DELETE':
+            return deleteGroup(request, name)
+    else:
+        return HttpResponse('{"errors":[{"message":"'+(_("You have not the rights to do this opperation").__unicode__())+'"}]}',content_type="application/json")
 
 
 def myself(request):
@@ -69,20 +79,29 @@ def myPassword(request):
 def categories(request):
     if request.method == 'GET':
         return showCategories()
-    if request.method == 'PUT':
-        return tablefactory.modifyCategories(request)
+    if request.user.admin :
+        if request.method == 'PUT':
+            return tablefactory.modifyCategories(request)
+    else:
+        return HttpResponse('{"errors":[{"message":"'+(_("You have not the rights to do this opperation").__unicode__())+'"}]}',content_type="application/json")
+
 
 
 def category(request, name):
-    if request.method == 'DELETE':
-        return tablefactory.deleteCategory(name)
+    if request.user.admin :
+        if request.method == 'DELETE':
+            return tablefactory.deleteCategory(name)
+    else:
+        return HttpResponse('{"errors":[{"message":"'+(_("You have not the rights to do this opperation").__unicode__())+'"}]}',content_type="application/json")
+
 
 
 def tables(request):
     if request.method == 'GET':
         return showAllTables(request.user)
     if request.method == 'POST':
-        return tablefactory.createTable(request)
+        if request.user.admin or request.user.tableCreator :
+            return tablefactory.createTable(request)
 
 
 def table(request, name):
@@ -119,7 +138,7 @@ def history(request, tableName):
     if request.method == 'GET':
         response = TableSerializer.serializeHistory(tableName)
         if not response:
-            return HttpResponse(json.dumps({"errors": [{"code": Error.TABLE_NOTFOUND, "message": "Could not find table with name " + tableName + "."}]}), content_type="application/json")
+            return HttpResponse(json.dumps({"errors": [{"code": Error.TABLE_NOTFOUND, "message": _("Could not find table with name ") + tableName + "."}]}), content_type="application/json")
         return HttpResponse(json.dumps(response), content_type="application/json")
 
 
@@ -140,7 +159,7 @@ def filterDatasets(request, tableName):
     if request.method == 'POST':
         datasets = DatasetSerializer.serializeBy(json.loads(request.raw_post_data), tableName, request.user)
         if datasets is None:
-            return HttpResponse(content=_("An error occured"), status=500)
+            return HttpResponse(json.dumps({"errors": [{"code": Error.TABLE_NOTFOUND, "message": _("An error occured")}]}), content_type="application/json")
         return HttpResponse(json.dumps(datasets), content_type="application/json")
 
 
@@ -155,7 +174,7 @@ def register(request):
     jsonRequest = json.loads(request.raw_post_data)
     try:
         DBUser.objects.get(username=jsonRequest["username"])
-        return HttpResponse(_("user with name ") + jsonRequest["username"] + _(" already exists."), status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("user with name ") + jsonRequest["username"] + _(" already exists.")}]}),content_type="application/json")
     except DBUser.DoesNotExist:
         user = DBUser.objects.create_user(username=jsonRequest["username"], password=jsonRequest["password"])
         user.save()
@@ -185,7 +204,7 @@ def showAllUsers():
 def showOneUser(name):
     user = UserSerializer.serializeOne(name)
     if user is None:
-        return HttpResponse(_("User does not exist"), status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": -1, "message": _("User does not exist")}]}), content_type="application/json")
     else:
         return HttpResponse(json.dumps(user), content_type="application/json")
 
@@ -203,7 +222,7 @@ def modifyUserRights(request):
         try:
             user = DBUser.objects.get(username=userObj["name"])
         except DBUser.DoesNotExist:
-            HttpResponse(_("Could not find user with name ") + userObj["name"] + ".", status=400)
+            HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Could not find user with name ") + userObj["name"] + "."}]}), content_type="application/json")
 
         if userObj["tableCreator"] != user.tableCreator\
                 or userObj["userManager"] != user.userManager\
@@ -215,8 +234,8 @@ def modifyUserRights(request):
         user.is_active = userObj["active"]
         if modified:
             user.save()
-
-    return HttpResponse(_("Successfully modified user rights."), status=200)
+        #TODO JSON
+    return HttpResponse(_("Successfully modified user rights."), content_type="application/json")
 
 
 def showAllGroups():
@@ -237,7 +256,7 @@ def addGroup(request):
         groupNames.append(name)
 
     if request["name"] in groupNames:
-        HttpResponse(content=_("Group with name ") + request["name"] + _(" already exists."), status=400)
+        HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Group with name ") + request["name"] + _(" already exists.")}]}), content_type="application/json")
 
     groupF = DBGroupForm({"name": request["name"]})
     if groupF.is_valid():
@@ -268,7 +287,8 @@ def addGroup(request):
             membership.save()
 
     if len(failed) > 0:
-        return HttpResponse({"error": _("following users could not be added to the group: ") + str(failed) + _(". Have you misspelled them?")}, content_type="application/json")
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message":  _("following users could not be added to the group: ") + str(failed) + _(". Have you misspelled them?")}]}), content_type="application/json")
+    #TODO JSON
     return HttpResponse(_("Successfully saved group ") + request["name"] + ".", status=200)
 
 
@@ -284,13 +304,13 @@ def modifyGroup(request, name):
     try:
         group = DBGroup.objects.get(name=name)
     except DBGroup.DoesNotExist:
-        return HttpResponse(content=_("Could not find group with name ") + name + ".", status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Could not find group with name ") + name + "."}]}), content_type="application/json")
 
     request = json.loads(request.raw_post_data)
     if request["name"] != group.name:
         try:
             DBGroup.objects.get(name=request["name"])
-            return HttpResponse(content=_("A group with name ") + request["name"] + _(" already exists."), status=400)
+            return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("A group with name ") + request["name"] + _(" already exists.")}]}), content_type="application/json")
         except DBGroup.DoesNotExist:
             group.name = request["name"]
 
@@ -334,7 +354,7 @@ def modifyGroup(request, name):
         theUser = DBUser.objects.get(username=admin)
         membership = Membership.objects.get(user=theUser)
         membership.delete()
-
+    #TODO JSON
     return HttpResponse(_("Successfully modifed group ") + name + ".", status=200)
 
 
@@ -342,11 +362,11 @@ def deleteGroup(request, name):
     try:
         group = DBGroup.objects.get(name=name)
     except DBGroup.DoesNotExist:
-        return HttpResponse(content=_("Could not find group with name ") + name + ".", status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Could not find group with name ") + name + "."}]}), content_type="application/json")
 
     Membership.objects.filter(group=group).delete()
     group.delete()
-
+    #TODO JSON
     return HttpResponse(content=_("Successfully deleted group ") + name + ".", status=200)
 
 
@@ -368,6 +388,7 @@ def changeMyPassword(request):
     jsonRequest = json.loads(request.raw_post_data)
     request.user.set_password(jsonRequest["password"])
     request.user.save()
+    #TODO JSON
     return HttpResponse(_("Saved password successfully."), status=200)
 
 
@@ -391,7 +412,7 @@ def showDatasets(request, tableName):
     try:
         Table.objects.get(name=tableName)
     except Table.DoesNotExist:
-        return HttpResponse(content=_("Could not find table with name ") + tableName + ".", status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Could not find table with name ") + tableName + "."}]}), content_type="application/json")
 
     jsonRequest = json.loads(request.raw_post_data)
     result = dict()
@@ -406,14 +427,14 @@ def showDataset(tableName, datasetID, user):
     try:
         table = Table.objects.get(name=tableName)
     except Table.DoesNotExist:
-        return HttpResponse(content=_("Table with name ") + tableName + _(" could not be found."), status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Table with name ") + tableName + _(" could not be found.")}]}), content_type="application/json")
     try:
         dataset = Dataset.objects.get(datasetID=datasetID, table=table)
     except Dataset.DoesNotExist:
-        return HttpResponse(content=_("dataset with id ") + datasetID + _(" could not be found in table ") + tableName + ".", status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("dataset with id ") + datasetID + _(" could not be found in table ") + tableName + "."}]}), content_type="application/json")
 
     if dataset.deleted:
-        return HttpResponse(_("The requested dataset does not exist."), status=400)
+        return HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("The requested dataset does not exist.")}]}),content_type="application/json")
     else:
         dataset = DatasetSerializer.serializeOne(datasetID, user)
         return HttpResponse(json.dumps(dataset), content_type="application/json")
@@ -422,7 +443,7 @@ def showDataset(tableName, datasetID, user):
 def showAllTables(user):
     tables = TableSerializer.serializeAll(user)
     return HttpResponse(json.dumps(tables), content_type="application/json") if tables is not None \
-        else HttpResponse(status=500)
+        else HttpResponse(json.dumps({"errors": [{"code": Error.USER_NOTFOUND, "message": _("Table could not be found")}]}),content_type="application/json")
 
 
 def showTable(name, user):
